shared::binary_helpers

Module le

Exposes helpers for working with raw binary data stored in little-endian format

Macros

impl_little_endian_integer 🔒

Traits

LittleEndianInteger

A trait for types that can be serialized and deserialized in little-endian format. This trait is implemented for integer types like u16, u32, and u64.

Functions

read_le

Reads a value of type T from a byte slice in little-endian format.

write_le

Writes a value of type T into a byte slice in little-endian format.