shared::schemas

Module schema

Defines Schema and related structs

Structs

Schema

An in memory representation of the schema of a table or index

SchemaBuilder

A builder of a Schema

SchemaEntry

A single entry in the Schema container

SchemaEntryId

Represents the unique identifier of a column within a SchemaEntry.