Module shared::structs::prelude

Re-exports

pub use super::fragment::Fragment;

pub use super::fragment_request::FragmentRequest;

pub use super::fragment_result::FragmentResult;

pub use super::fragment_task::FragmentTask;

pub use super::pixel_data::PixelData;

pub use super::pixel_intensity::PixelIntensity;

pub use super::point::Point;

pub use super::range::Range;

pub use super::resolution::Resolution;

pub use super::u8data::U8Data;

pub use super::fractals::iterated_sin_z::IteratedSinZ;

pub use super::fractals::julia::Julia;

pub use super::fractals::mandelbrot::Mandelbrot;

pub use super::fractals::newton_raphson_z3::NewtonRaphsonZ3;

pub use super::fractals::newton_raphson_z4::NewtonRaphsonZ4;

pub use super::fractals::*;

pub use super::fractal_descriptor::*;

Structs

Complex

Traits

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

Serialize

A data structure that can be serialized into any data format supported by Serde.

Derive Macros

Deserialize

Serialize